# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VortexUI Panel API
  version: "0.1.0"
  description: >
    Control-plane REST API for VortexUI. All endpoints under `/api` (except
    `/api/login` and `/api/health`) require a Bearer JWT obtained from
    `/api/login`. Mutating routes additionally require an RBAC permission
    (sudo admins bypass all permission checks). The public `/sub/{token}`
    endpoint is authenticated solely by the opaque subscription token.
servers:
  - url: http://localhost:8080
    description: Local panel
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message: {type: string}
...
    NodeHealth:
      type: object
      properties:
        cpu_percent: {type: number}
        mem_percent: {type: number}
        disk_percent: {type: number}
        core_running: {type: boolean}
        connections: {type: integer}
    Node:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        address: {type: string, example: "1.2.3.4:50051"}
        core: {type: string, enum: [xray, singbox]}
        status: {type: string, enum: [connected, disconnected, error, disabled]}
        usage_ratio: {type: number}
        last_seen: {type: string, format: date-time, nullable: true}
        health: {$ref: '#/components/schemas/NodeHealth'}
        core_version: {type: string}
        agent_version: {type: string}
        created_at: {type: string, format: date-time}
...
  responses:
...
    NotFound: {description: not found, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
security:
  - bearerAuth: []
paths:
...
  /api/nodes:
    get:
      tags: [Nodes]
      summary: List nodes (perm node:read)
      responses:
        "200": {description: nodes, content: {application/json: {schema: {type: object, properties: {nodes: {type: array, items: {$ref: '#/components/schemas/Node'}}}}}}}
    post:
      tags: [Nodes]
      summary: Create + register a node (perm node:write)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address]
              properties:
                name: {type: string}
                address: {type: string}
                core: {type: string, enum: [xray, singbox]}
                usage_ratio: {type: number}
      responses:
        "201": {description: created, content: {application/json: {schema: {type: object, properties: {node: {$ref: '#/components/schemas/Node'}, warning: {type: string}}}}}}
  /api/nodes/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
    get:
      tags: [Nodes]
      summary: Get a node (perm node:read)
      responses:
        "200": {description: node, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
        "404": {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Nodes]
      summary: Update a node, re-establishing its connection (perm node:write)
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, properties: {name: {type: string}, address: {type: string}, usage_ratio: {type: number}}}}}
      responses:
        "200": {description: updated, content: {application/json: {schema: {type: object, properties: {node: {$ref: '#/components/schemas/Node'}, warning: {type: string}}}}}}
        "404": {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Nodes]
      summary: Deregister + delete a node (perm node:write)
      responses:
        "204": {description: deleted}
...