# internal/platform/postgres/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Admin struct {
	ID           uuid.UUID
	Username     string
	PasswordHash string
	Sudo         bool
	RoleID       pgtype.UUID
	TotpSecret   string
	TotpEnabled  bool
	UserQuota    int32
	TrafficQuota int64
	LastLogin    pgtype.Timestamptz
	CreatedAt    pgtype.Timestamptz
}

type Balancer struct {
	ID            uuid.UUID
	NodeID        uuid.UUID
	Tag           string
	Selectors     []byte
	Strategy      string
	Observe       bool
	ProbeUrl      string
	ProbeInterval string
	Enabled       bool
}

type Inbound struct {
	ID               uuid.UUID
	NodeID           uuid.UUID
	Tag              string
	Protocol         string
	Listen           string
	Port             int32
	Network          string
	Security         string
	Sni              []byte
	Path             string
	Host             []byte
	Flow             string
	EvasionProfileID pgtype.UUID
	Raw              []byte
	Enabled          bool
}

type Node struct {
	ID           uuid.UUID
	Name         string
	Address      string
	Core         string
	Status       string
	UsageRatio   float64
	LastSeen     pgtype.Timestamptz
	CpuPercent   float64
	MemPercent   float64
	DiskPercent  float64
	CoreRunning  bool
	Connections  int32
	CoreVersion  string
	AgentVersion string
	CreatedAt    pgtype.Timestamptz
}

type Outbound struct {
	ID       uuid.UUID
	NodeID   uuid.UUID
	Tag      string
	Protocol string
	Address  string
	Port     int32
	Uuid     string
	Password string
	Username string
	Method   string
	Flow     string
	Network  string
	Security string
	Sni      string
	Path     string
	Host     string
	Raw      []byte
	Enabled  bool
}

type Role struct {
	ID          uuid.UUID
	Name        string
	Permissions []byte
}

type RoutingRule struct {
	ID          uuid.UUID
	NodeID      uuid.UUID
	Priority    int32
	Name        string
	InboundTags []byte
	Domains     []byte
	Ip          []byte
	Port        string
	Protocols   []byte
	Network     string
	OutboundTag string
	BalancerTag string
	Enabled     bool
}

type TrafficPoint struct {
	Time   pgtype.Timestamptz
	UserID uuid.UUID
	NodeID uuid.UUID
	Up     int64
	Down   int64
}

type User struct {
	ID            uuid.UUID
	Username      string
	Status        string
	Note          string
	DataLimit     int64
	UsedTraffic   int64
	ExpireAt      pgtype.Timestamptz
	OnHoldExpire  pgtype.Int8
	ResetStrategy string
	LastReset     pgtype.Timestamptz
	DeviceLimit   int32
	AllowedHwids  []byte
	VmessUuid     uuid.UUID
	VlessUuid     uuid.UUID
	TrojanPass    string
	SsPassword    string
	SsMethod      string
	SubToken      string
	CreatedAt     pgtype.Timestamptz
	UpdatedAt     pgtype.Timestamptz
}

type UserInbound struct {
	UserID    uuid.UUID
	InboundID uuid.UUID
}